+++ w8ing/__version__.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 """
 MIT License
 
@@ -28,4 +29,4 @@ SOFTWARE.
 ╚███╔███╔╝╚█████╔╝██║██║ ╚████║╚██████╔╝
  ╚══╝╚══╝  ╚════╝ ╚═╝╚═╝  ╚═══╝ ╚═════╝
 """
-__version__ = "1.0.0"  # pragma: no cover
+__version__ = "1.0.1"  # pragma: no cover
